Remove unused variables
